+++ bevy_rapier2d/examples/testbed2.rs
@@ -147,7 +147,8 @@ fn main() {
             OnExit(Examples::PlayerMovement2),
             (
                 cleanup,
-                |mut rapier_config: ResMut<RapierConfiguration>, ctxt: Res<RapierContext>| {
+                |mut rapier_config: ResMut<RapierConfiguration>,
+                 ctxt: Query<&mut RapierContext>| {
                     rapier_config.gravity =
                         RapierConfiguration::new(ctxt.integration_parameters.length_unit).gravity;
                 },