+++ bevy_rapier3d/examples/ray_casting3.rs
@@ -76,10 +76,11 @@ pub fn setup_physics(mut commands: Commands) {
 pub fn cast_ray(
     mut commands: Commands,
     windows: Query<&Window, With<PrimaryWindow>>,
-    rapier_context: Res<RapierContext>,
+    mut rapier_context: Query<&mut RapierContext>,
     cameras: Query<(&Camera, &GlobalTransform)>,
 ) {
     let window = windows.single();
+    let rapier_context = &mut *rapier_context.single_mut();
 
     let Some(cursor_position) = window.cursor_position() else {
         return;